refactor(components) : categorize objects
